apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
aioredis==2.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
google-generativeai==0.3.1
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
python-json-logger==2.0.7
aiofiles==23.2.1
bullmq==1.0.0
websockets==12.0
tenacity==8.2.3
